+++ src/renderer/components/swap/EditableAddress.styles.tsx
